+++ src/main/java/net/itsky/java/sort/metric/MetricDataCyrForestCached.java
@@ -8,19 +8,19 @@ import java.util.concurrent.ConcurrentMap;
 
 public class MetricDataCyrForestCached extends MetricDataCyrForest {
 
-    private ConcurrentMap<String, Long> cache = new ConcurrentHashMap<>();
+    private ConcurrentMap<String, Integer> cache = new ConcurrentHashMap<>();
 
     @Override
-    public long metric(String s) {
+    public int metric(String s) {
         if (s == null || s.isEmpty()) {
             return super.metric(s);
         }
         String key = s.length() <= 2 ? s : s.substring(0, 2);
-        Long val = cache.get(key);
+        Integer val = cache.get(key);
         if (val == null) {
             val = super.metric(key);
             cache.put(key, val);
         }
-        return val.longValue();
+        return val.intValue();
     }
 }
